feature/PBAO-243-Move-all-encryption-infrastructure-classes-into-EasyEncryption
- Combine DoctrinePostLoadEncryptionListener and DoctrineOnFlushEncryptionListener into DoctrineEncryptionListener
